SimpCfg+DataUtilsString: Move string helpers to its own file
